Update patch set 7
Patch Set 7:

Kaldari:
> Isn't using cookies better than failing to store the data??

Maybe, maybe not. It depends how important it is.

It seems the main proposed reason for having cookies is in case localStorage runs out of space (are there any browsers MobileFrontend actually supports that don't support any of the other non-cookie, non-memory backends?)

But it seems that a better solution to the space problem may be expiring old stored data (if they run out of local/session/globalStorage/userData space, we don't want to use cookies forevermore).

With localStorage, it seems expiration has to be implemented at the application layer.  This is part of that extra code jStorage has (look for TTL there).

Patch-set: 7