Update patch set 7
Patch Set 7: Code-Review+1

This looks fine - although a follow up patch showing it in use would be useful, as I worry without that we would end up 

It seems like Ori should weigh in a bit more before +2ing.

Our existing use of local storage is purely convenience so in answer to @kaldari failure to store data is not the end of the world

Our current use cases are:
* to store the users preferred language
* to decide whether sections are expanded by default or not
* to set a preferred editor
* to aid click tracking on links by delaying it till the next page view
* to cache nearby search results

Out of those the only real annoying one is we might lose click tracking on devices with low local storage - but this could be worked around by adding a test to see if localStorage is full. None of these use cases are essential to the mobile experience.

A better fallback might even be to use user preferences and make them work across devices...

Patch-set: 7