Update patch set 7
Patch Set 7:

Right now, I run out of localStorage space on a regular basis when using Wikipedia. It's quite annoying having to manually clean out my localStorage in order to have it remember my anonymous settings. I don't see how this is a false dilemma. LocalStorage space is limited, cookie space is not (well technically it is, but not in a way that you have to worry about).

Regarding other storage options, none of the other options are persistent besides cookies.

What if I changed the default cookie expiration from 1 year to 1 month? At least then it is unlikely that the user's browser would fill up with too many cookies.

At this point I feel like I'm being persecuted by religious fundamentalists. Cookies suck, but they work, and we've been using them (conservatively) since the project began. I still have yet to hear a convincing argument why using them as a fallback is a bad idea. This patch isn't to fix a hypothetical situation, it's to make using Wikipedia not suck for me personally. If anyone has an alternative solution that would actually work (and be persistent for anonymous users), please let me know. The idea of making localStorage data expire sounds promising, but I wonder how well we could enforce that as a best practice (and whether or not it would actually solve the problem of using up localStorage space).

Patch-set: 7