Update patch set 7
Patch Set 7:

> Right now, I run out of localStorage space on a regular basis when using Wikipedia. It's quite annoying having to manually clean out my localStorage in order to have it remember my anonymous settings.

Desktop or mobile, with what browser (they have different storage limits, so it's important to see what we're dealing with)?  Also, what settings can't be saved when it's full?

Checking what's actually in the storage would be helpful.

> I don't see how this is a false dilemma. LocalStorage space is limited, cookie space is not (well technically it is, but not in a way that you have to worry about).

Individual cookies are limited to 4K.  http://stackoverflow.com/a/10889034/47773 says Mobile Safari and Android support 50 per host, which is 200K.  That's a far smaller limit than localStorage, which is 5 MB on iPhone according to http://stackoverflow.com/a/2679277/47773 , and similar in other browsers from my understanding.

The bandwidth issue is also significant, and more important than the absolute maximum limit.

> Regarding other storage options, none of the other options are persistent besides cookies.

This is not true.  globalStorage (https://developer.mozilla.org/en-US/docs/Web/Guide/API/DOM/Storage) and userData are fully persistent, sessionStorage is semi-persistent (survives page restore, etc.)

> What if I changed the default cookie expiration from 1 year to 1 month? At least then it is unlikely that the user's browser would fill up with too many cookies.

It's unclear if this will make a difference in practice.  It depends how often the cookie is re-saved.

> I still have yet to hear a convincing argument why using them as a fallback is a bad idea.

I'm more supportive of using them as a fallback for particular important settings than as a general catch-all.

> This patch isn't to fix a hypothetical situation, it's to make using Wikipedia not suck for me personally.

As requested above, please explain exactly what sucks in the current situation (which settings can't you save, what's filling it up, etc.).

> The idea of making localStorage data expire sounds promising, but I wonder how well we could enforce that as a best practice (and whether or not it would actually solve the problem of using up localStorage space).

jStorage already has this (of course, any proposed new library could do the same), so it's just a question of code review to make sure people using use options.expires.  It handles the rest behind the scenes.

Patch-set: 7