Update patch set 6
Patch Set 6: Code-Review-1

(1 comment)

Breaks the header in search overlay slightly. Type a search term, the X (clear input) button is too far to the right. Also, the right margin for the bottom border is missing, which leads me to think that the entire header is too wide.

Patch-set: 6
Label: Code-Review=-1