Update patch set 1
Patch Set 1:

This icon is from the dropbox of icons that May has been maintaining. It's also the one in the Winter prototype.

I have mentioned it to Kaity/Kenan in the past and she wanted to use this icon and it was supposed to be pushed when we pushed search to stable, but I think it fell of the radar as it seems non-JS site is way down the priority list. I ran this by Kaity another 2 weeks ago but it never got prioritised.

Remember non-JavaScript users include anyone we blacklist on RL - which is pretty much anyone not on Firefox, Android or iPhone.

Anyway... have you seen the current non-JavaScript button!? :) This is definitely in my opinion an improvement on the current situation... feel free to run by a designer if you like but I reckon they have bigger fish to fry and we should just merge this. The icon particularly looks ugly on a tablet/desktop (although I'm not sure how many this will impact better but I feel it is better to be safer than sorry)

I've removed the path file.

Patch-set: 1
Label: Code-Review=0