Removing header margin and associated hacks
This is necessary in order for the VE dropdowns to render the same
in Firefox and Chrome. Otherwise their vertical positions are
different.

Also changing @headerTitleMarginH to @headerTitlePaddingH so it is
accurately named.

Also removes vertical lines to sides of buttons at request of Kaity.

Also making bell icon slightly smaller per Maryana and Kaity.

Change-Id: I879cc6096d5c0481b5d8194c0defa5cbce8d9f6d