Promoting mobile typography changes from beta to stable
This change promotes the updates to tablet styles to stable. The
way these styles is loaded has been changed to use a regular top
position ResourceLoader module rather than the fancy matchMedia
hack, as it seems a bit risky for stable. The tablet styles are
pretty small anyway, so it doesn't seem like it would be a
significant change to the footprint of the head.

Change-Id: Ia2ddca666c42762378e4c2bfe8e6a82ee3e43420