Update patch set 1
Patch Set 1:

but it's not zero...?
If a page hasn't had any activity for a while and hasn't been edited recently it will be 0 rather than the actual number of top level sections.

We should make a call on what to do with this code. It's been sitting in alpha for a while and we should decide whether to abandon it or move it to beta.

Patch-set: 1
Label: Code-Review=0