Update patch set 2
Patch Set 2:

Yeah, you are right. But, composer.lock is in git (correctly) to install all dependencies correct. But with this change https://gerrit.wikimedia.org/r/#/c/135083/ all scripts directory was renamed to dev-script to avoid confusing. But this is exactly that, what was :) The scripts/phpcs and scripts/phpcbf in bin array are for composer to know, which symlinks has to generated in vendor/bin. Unhappily MaxSem has changed this to dev-scripts, which does not exist in squizlabs/php_codesniffer, thats the reason for the error message i wrote in E-Mail (skipping bc of not exist). So, after the chnage back to scripts/phpcs and scripts/phpcbf (which exist in squizlabs/php_codesniffer) the bin directory and symlinks can created correctly. Hope, that this was understandable :)

Patch-set: 2