Update patch set 3
Patch Set 3:

Juliusz: I suspect this is the change in pagelist.less but I cannot replicate so would appreciate if you can point me at the bug and post a screenshot.

The FIXME I remove for line-height: normal mentions remove when go to stable and I suspect this is what is causing the issue but I cannot verify. If this is the culprit we need to update that comment so we know what it is therefore.

Patch-set: 3
Label: Verified=0