Update patch set 3
Patch Set 3: Code-Review-1

This looks much better! yey! :)
That said this works fine in beta but not in stable.

Property: SkinMinerva::$echoButtonClass in /Users/jrobson/Sites/w/extensions/MobileFrontend/includes/skins/SkinMinerva.php on line 25

Also in stable mode I simply see a '1' e.g. no alarm icon.

I would consider dropping the number altogether as it is not true representation of talk page. Simply show the alarm icon when there are messages and don't when there are not.

Patch-set: 3