Update patch set 1
Abandoned

It was a proof of concept. Feel free to resurrect it later. I still would like to hear more about this meeting though if there was more than what was in the e-mail, as according to the latest e-mail this might be a good enough approach.

Patch-set: 1
Status: abandoned