Update patch set 4
Patch Set 4: Code-Review-1

I'm a bit skeptical of the approach here. Isn't there a chance that this will cause some events to be lost if there is a page load in the chain after initialization? Is chaining an event handler for an initialization event off the initialization itself really a pattern we need to support? In cases like that shouldn't we just refactor the code so that the initialization performs the action directly?

Patch-set: 4
Reviewer: Kaldari <78@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1