Update patch set 3
Patch Set 3: Code-Review-1

(1 comment)

Jenkins says unit tests are failing.

Max, I'm nervous about this patch - have you tested this in an environment with a separate mobile domain? IIRC, that's where the original problem manifested. I believe the behavior was as follows: toggle from desktop view to mobile view and back to desktop view (to ensure 'stopMobileRedirect' cookie gets set). then, toggle back to mobile view - iirc, what happened in this case was that you'd see the mobile view but the cookie would NOT get unset (because the page was loaded from browser cache). this was a long time ago and i dont think there was a bug opened for it, so i could be off with the specifics, but that was at least the general idea of what was going on.

Patch-set: 3
Label: Code-Review=-1