Update patch set 2
Patch Set 2: Code-Review-1

Looks like the solution is going to have to be a bit more complicated since the toc.js code actually requires the toggling code (so that you can expand and collapse the TOC). Either toc.js should be rewritten to not require the toggling code, or the init function in toggle.js should be smarter about where it is executing.

In the meantime, feel free to split the bug 67105 stuff into a separate commit, and I can go ahead and merge this one so that bug 67101 is fixed.

Patch-set: 2
Label: Code-Review=-1