Update patch set 2
Patch Set 2: Code-Review-1

This seems a but redundant with the existing functionality. Currently, the upload page-action button is only displayed if the user is autoconfirmed, which includes a minimum edit threshold. What we might want to do instead is to extend this restriction to the upload button in the sidebar menu as well, since about half our uploads come from there.

Patch-set: 2
Label: Code-Review=-1