Update patch set 1
Patch Set 1:

The ideal would be to protect the page via the API.  Doing it in the UI is clumsy and expensive.

This patch has the benefit of testing for protecting the page in one test, and visiting that page in another test.

Another solution would be to use a dedicated user not Selenium_user for this test.

Regardless, the more parallel we run tests, the more an explicit logout for Selenium_user is going to cause problems for other tests.

Patch-set: 1
Reviewer: Cmcmahon <77@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0