Hygiene: Cleaning up pageaction icon CSS
Removing overly-specific CSS positioning for pageaction icons that
assumes the icons always appear in a certain order. This is a bad
assumption as we often hide certain icons under various conditions.
Replacing the per-icon positioning with a generic rules that
always put 16px right-margin for mobile view and 16px left margin
for tablet view.

Change-Id: I375014103f8d1e208861e567f3879b3c94becbd6