Update patch set 16
Patch Set 16: Code-Review-1

returnto parameter gets set incorrectly.

Also minor tweak to UI - i'd put the edit anonymously link at the bottom underneath the signup. The signup and login links are associated with one another, so putting an edit anonymously in between is a little confusing.

That aside I think this is great and once these issues are fixed we should give it ago in alpha and see what happens! :-)

http://localhost:8080/wiki/Special:UserLogin?returnto=Foo%23undefined&campaign=mobile_editPageActionCta&returntoquery=article_action%3Dsignup-edit&type=signup

Patch-set: 16