Improvements for viewing mobile site on very small screens (<=280px)
Hide the pageactions bar for devices with small screens and no
JavaScript support.

Add a border around the search field so that it is clear where to
click in Opera Mini (which doesn't support placeholders).

Also improving the formatting of the search field and adjacent
search icon so that they don't look weird with the new border.

Change-Id: I9ce307ad687a376596cebb9088ef49c3738f9823