Update patch set 2
Patch Set 2:

Given that stable only shows the Watchlist icon for devices at the intersection of no/RL-impaired JS && small screen size, this works. Other JS features ought to be hidden in page-actions as well, but for now I think only in Beta mode is that an issue at the moment (https://bugzilla.wikimedia.org/show_bug.cgi?id=67819), so that can be addressed in a follow-up patch.

Patch-set: 2