Making sure that clicking on search doesn't trigger history.back()
As part of this, I'm removing the weird setTimeout on the router
navigation for search. I've tested this in Mobile Safari, Chrome 35
on Android, and the native Android browser (4.1.1), and didn't have
any problems with the keyboard. Perhaps it was a bug in an old
version of Chrome.

Also adding some more comments to SearchOverlay.js.

Bug: 67140
Change-Id: If6149cb5b8236c6d701a8441ac85360ba06240b0