Update patch set 10
Patch Set 10: Code-Review-1

Sorry, i misunderstand you :) I thought you mean, when you manually write it (don't know how i came to this thinking).

But i think i found a problem, but it's late, so maybe the problem is before the computer ;) But in the actual change there wouldn't be selected the right section line? The first section line is that one we want to have, but we use the n-one (n -> id of which section we want to edit). So as example we edit section 3 we will get the section titles as an array from preview api request. The section title of section 3 is there with index 0, but we will use the section title with index 3, what's the false one :/

Patch-set: 10
Label: Code-Review=-1