Update patch set 3
Patch Set 3: Code-Review-1

This would be a better way to do this, but will not work well (see the comments in the preferences hook). The current system in core is kind of incompatible with the beta feature magic.

Patch-set: 3
Reviewer: Bartosz Dziewoński <417@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1