Update patch set 1
Patch Set 1: Code-Review-1

Actually, it is not THAT simple.

Although legally allowed to have two roles, probably there is not a single browser that actually interprets anything but the first role.

As such it should probably be role="heading button", emphasizing the primary and implicit role. The tabIndex=0 will probably make it actionable regardless of wether or not the browser understands the 2nd button role. Alternatively this role can be removed altogether (if we forego progressive enhancement).

But looking even further into the page source, there is something more serious going on I would say that is causing this problem. The usage of h2 to wrap the headline and the edit button, which strictly speaking is incorrect (also a problem in the main website btw and there is a ticket for it), and partially explains why this hack of "heading button" is required at all.

The whole structure of the page should really get some changes. It should be more like:
<section>
<header class="section_heading><h2 class="mw-headline"><a class="edit-page>Edit</a></header>
<div class="content_block"><!-- this div is needed for the collapsing -->
content
</div>
</section>

Patch-set: 1