Update patch set 1
Patch Set 1:

Sure, but I'm not doubting the choice to use js here. That's fine. But I noticed you have a tendency to use html() in this way which looks like you might not be aware of what that actually does. So be wary of that in the future, and consider simply moving the nodes via getting the detached contents and appending to the new span instead of creating a copy by cloning or serialising/parsing them.

Patch-set: 1