Update patch set 3
Patch Set 3: Code-Review-1

Okay so I think the problem you are having is you are using $ rather than this.$

The overlay has not been added to the DOM and when you do $( '.spinner' ) or $( '.page-list.actionable' ) is will not match any elements.

Switch to using this.$( '.page-list.actionable' )
Basically what this does is search the overlay element (which has not been added to the DOM for the element requested)

I think this is where your problems lie.

Patch-set: 3