Update patch set 2
Patch Set 2:

Page content is always supposed to have a 16px margin on phones (small screens). This makes it visually clear that there are 3 separate areas of the page: header, content, and footer (as the only elements going all the way to the edges are the horizontal rules separating those 3 areas). I don't see why the uploads page should be an exception to that.

Patch-set: 2
Reviewer: Kaldari <78@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0