Update patch set 3
Patch Set 3: Code-Review-1

@Jon: This problem isn't fixed in this change (it's still in master), but fixed with the change to imrpove the add discussion workflow (i have no change id in clipboard, but it's the last change of this dependency tree).

But here is another problem:
If the title of a discussion is longer as the width of the display it should be shorted like in editor (Edit page Test...). And there is the question: is there a need to show the full doscussion name in the overlay itself (like it is in master actual)? Normally the user see the title in the overbiew.

Patch-set: 3
Reviewer: Florianschmidtwelzow <890@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1