Update patch set 14
Patch Set 14:

> Actual: loads talk overlay then closes it

Hmm, it seems, that it takes a long time to remove the overlay div from the DOM :/ That is the reason, that it will be removed when you switch with back and forward key's of the browser in "normal" speed. If you wait a few seconds it works like expected. That needs more work :)

Patch-set: 14