Update patch set 4
Patch Set 4:

Hmm, right. But message parsing with JS doesn't work with {{PAGENAME}}, like used in de.wikipedia, what about pageApi? Load the content of MediaWiki:Anoneditwarning only when really used (user is anon AND clicked edit without logging in). That sounds to be a better solution for this.

Patch-set: 4