Update patch set 9
Patch Set 9: Code-Review-1

The behaviour is incorrect. This may be related to the prior patch.

Creating a topic results in seeing an overlay with "unnamed discussion" at top and a loader below and a toast saying my topic was added (but no visual feedback). The loader clears, the overlay closes and then reopens.

Expected: the ajax loader should disappear and be replaced by the new content. The overlay shouldn't close at all under any circumstances.

Patch-set: 9
Label: Code-Review=-1