Update patch set 10
Patch Set 10: Code-Review+1

I chatted with Max in IRC - he mentioned that 'mobileaction' in the query string will result in the request missing the Varnish cache on the WMF cluster. So I think this is all OK; and this appears to work as expected for me locally (on a single-domain configuration).

I would prefer it if someone else from the mobile team do the final +2 here - it's time to pass the torch on MobileFrontend cookie handling knowledge and +2ability ;) I've added kaldari as a reviewer.

Patch-set: 10
Reviewer: Awjrichards <91@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1