Update patch set 6
Patch Set 6:

I think this isn't the best way handling this. First transfer the entire message as wikitext to the client (load with module), then transfer it back to the api interface (as text in api) to get back the entire message as parsed html back from the api. That sounds terribly inefficient.

Another idea is to extend the mobile API with a function to get a parsed message. This would reduce the data transfer to 1/3 of this change. And it would be possibly a good way to handle the bug with the summary message (bug 69488). The only point i wondering: that sounds like a core feature?

Patch-set: 6