Update patch set 26
Patch Set 26:

> Can you rebase this Florian?
Sure, done :)

> I'm keen to wrap up this stuff this coming week!
I'm happy about every help and hint which helps to figure out the problem here. I'm not sure, if this is the right implementation of a route in OverlayManager, so i think it would be good to check this first :) I don't see, why the DOM is updated with this delay :( But for now: have a nice weekend! :)

Patch-set: 26