Escape all CSS metacharacters for ref drawer
Expand the regular expression in getReference() to include almost
all CSS meta characters rather than doing so in a piecemeal fashion.

Also switch to .find() for extracting the list elements rather than
string concatenation. The intent of the function is now more clear.

Move call to .html() out of getReference to simplify the
responsibility of that function and bring it in line with
the other extracted drawer property.

Bug: 70921
Change-Id: Ica6a00846455a0d75ffc57f40f2ed64408fbadc7