Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

Giving a -1 over removing the click and the minor lack of space (see comment). My reservations over the refactoring are secondary. :)

I'm not sure I understand how the previous implementation breaks the PO pattern—there's still a clear separation of concern between performing the page action and the expectation of what the resulting state should be.

It also seems clearer to me to encapsulate the idiosyncrasies of what successful "typing into" involves across browsers (idiosyncrasies like clicking first, which we might still need for Chrome IIRC, and including a space). Clarity is subjective of course, as is the impetus to encapsulate or not to encapsulate—I'm definitely heavy on the abstraction at times.

Patch-set: 2
Label: Code-Review=-1