Update patch set 3
Patch Set 3:

I shouldn't have said Page Object, I changed the commit message. 

What I was after is that before when this test had

Then(/^I see the edit reflected in the article content$/) do
  on(ArticlePage) do |page|
    page.wait_until { page.toast.include?("Your edit was saved") }
    page.wait_until { page.content_element.visible? }

    expect(page.content).to include(@article_text)
  end
end

those "wait_until" statements should go in a "When" step and not in a "Then" step, because they are not RSpec assertions. 

Coincidentally,  I think we're getting "stale" objects on occasion because each time we call "on(ArticlePage)" we get a new object, and this can cause a race condition for Selenium.  I changed the test so that all the "When" steps are called on the same ArticlePage object, and only then a single assertion for the "Then" step, when the page should be in its final state.

I fixed the space thing too, I hadn't noticed that.

Patch-set: 3
Label: Verified=0