Update patch set 2
Patch Set 2: Code-Review-2

We already show
"A watchlist helps you bookmark pages and keep track of changes to them.
Log in to see it."

With your change this message is repeated.

What would be useful is for the top message to be generated using the warning query string parameter just like core as at some point we want to remove our customisations to the login form.

Patch-set: 2