Update patch set 5
Patch Set 5: Code-Review+1

The code looks fine but I don't really understand this bug still and why it is happening. Is it browser specific, is there any links you can find that describe the problem this fixes?

I worry if we look at this code in 3 months time we will say "what is this code for?"

Patch-set: 5