Update patch set 1
Patch Set 1: Code-Review+2

Seems to fix the loader in all the places I can think of it being invoked.
The loader on Wikigrok is still a bit funky though. It appears to close to the right hand side of screen. Since that's in alpha I guess that can be done as a follow up.

Patch-set: 1
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2