Update patch set 12
Patch Set 12: Code-Review-1

(3 comments)

A couple of minors inline. Next up: testing!

This is a very large change and so it's quite hard to keep this all in my head. Generally speaking it looks okay. The sooner this is tested, merged, and released, the sooner we can get rid of them there feature flags.

Maybe you could have broken this up into a sequence of changes: move the existing icon markup to be mw-ui-icon-esque; use the mw-ui-icon stylesheet if the user's on the alpha site.

Patch-set: 12
Label: Code-Review=-1