Update patch set 3
Patch Set 3:

Yeh there are a bunch of issues with the new icons that I was hoping to resolve in alpha, which of these do you think are you the most pressing to fix?

A lot of the tiny icons problems will be resolved by having a canonical icon library.

Patch-set: 3
Label: Verified=0