Update patch set 2
Patch Set 2:

Thanks, Jhernandez, the problem I'm having is that this will save the scroll position and restore it whenever the user visits the Nearby page. I think it should only restore the position when the user hits the browser's back button to view the Nearby page. I'm not sure how to do that.

Patch-set: 2