Update patch set 1
Patch Set 1: Code-Review-1

This logging needs to be split between the client and the server to minimise the number of requests that the client is making, i.e. if the server fails to fetch claim suggestions, then it can log that rather than the client.

Also, think about the conditions under which the getSuggestions/recordClaims APIs can fail, do you really want to try an log an error if the user's device doesn't have a connection?

Patch-set: 1
Label: Code-Review=-1