Fix search input and search button missalignment (js disabled)
There is a screenshot of the bug in the bug report.

With the fix everything is aligned, but it is very hard to keep the search icon aligned to
the right without over complicating it. I think it is good enough.

I've posted a couple of screenshots on the trello card of how it looks.
More discussion at: https://trello.com/c/guJYIOhs

Bug: 72637
Change-Id: I3abef85af15208cb966a8ab5bdd541b7c6000cf6