Update patch set 7
Patch Set 7: Code-Review-1

http://localhost:8080/wiki/Anne%20Dallas%20Dudley?mobileaction=alpha&wikidataid=Q3 is still giving me an infinite spinner when I click yes. We shouldn't introduce this bug.

Check with Moiz/Maryana what we should do in this case.

Currently dialog A does an ajax request before which is also bad as it is not scalable. We should fix the problem for both dialogs but at the very least fix it for B.

If nothing else we should at least clear the spinner and give the user some feedback. The infinite spinner looks unprofessional.

Apart from this, the code looks good to go.

Patch-set: 7
Label: Code-Review=-1