Update patch set 2
Patch Set 2:

Phuedx, I thought about your comment that we should log the errors in the backend, but it's still not as reliable as logging on the front end. The reason is that the even though the backend may not have errors and for some reason the front end is not able to get response from the backend, then the backend will not log an error. But the front end should.

Also, about no internet connection. If there is no connection, then logging won't work too.

What do you think? Should we just accept front end logging for now?

Patch-set: 2
Reviewer: Bmansurov <1963@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0