Update patch set 3
Patch Set 3: Code-Review-1

It does what it says, but it seems dumb to me and I don't get why we need to do this
.
Shouldn't we just write tests to ensure that recordClaims gets called appropriately. We control the code we write so why do we need this - shouldn't we just make sure we write the code correctly and write tests correctly? We shouldn't punish end users with more JS just because we are writing code badly..

Patch-set: 3
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1