Update patch set 3
Patch Set 3:

Sorry, this was a bad game of telephone. I asked Sam if he could add some basic error handling to the WikiGrokResponse API (on the server side). He then created a card for it, but didn't specify that it was referring to the server-side API, not the client-side API interface. Probably should have gone into Needs Analysis instead of Ready for Dev so that either me or Maryana could have checked the description and Acceptance Criteria. Probably I should have just created a card for it in the first place, instead of asking Sam informally :P

Patch-set: 3
Reviewer: Kaldari <78@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0