Update patch set 4
Patch Set 4: Code-Review-1

(4 comments)

> At a higher level though, I'm concerned about the API call. Wouldn't it be cheaper to embed this data onto the page as it's being rendered, say when the onResourceLoaderGetConfigVars hook fires?

Arghhhhhhhhhhhhh! As i read this, I remember the following line:
mw.config.get( 'wgCategories' );

The categories of a page (names without namespace) are already added to the source code of the page and can be used. So there is, at all, no need for an api call.

Patch-set: 4
Label: Code-Review=-1