Update patch set 1
Patch Set 1: -Code-Review

Yeh it seemed a little strange to me. Technically the languages list is not a next step in the process though. In search clicking the search icon applies the search term in this case it's just a button so it's probably alright. Neutral sadly has no usage instructions. https://bugzilla.wikimedia.org/show_bug.cgi?id=73252

Patch-set: 1