Update patch set 2
Patch Set 2: Code-Review-1

This will cause the logo to be rendered at full size without retina support. What we probably want to do is set the height to something like 15px (so it fits on the line) and set the width to $imageSize[0]/($imageSize[1]/15), or some fancy math like that. I would also suggest that we get a new version of the logo that is exactly twice the rendering height and width, i.e. 30px tall.

Patch-set: 2
Label: Code-Review=-1