Update patch set 1
Patch Set 1:

The way the qunit task is set up here requires a full MediaWiki install. Since this is not (and should not be) done as part of the 'npm install', you should not include 'qunit' in the 'npm test' command. The contract for 'npm test' is that it only requires a git clone (with submodules maybe) and 'npm install' to have run.

Any environment variables or applications installed prior should be done in a separate job. You'll find that qunit fails. Instead limit this one to static tests and builds for the moment, and use the mwext-qunit job for the qunit tests.

Patch-set: 1