Update patch set 1
Patch Set 1: Code-Review-1

"Given it's disabled except for users who opt-in, this will be
very low impact" – that's not entirely accurate. All users who see the edit screen will now see the switcher cog. This makes the title of the page very hard to read and takes up a lot of real-estate on a small and already nav-heavy screen. And given the prominence it now assumes (it's right next to save), the cog seems like the secondary call to action on the page and will probably see a lot of exploratory taps. 

I don't think this is ready for the full mobile site yet. As Jon said, it's fine to put this in alpha/beta and iterate on it, but we shouldn't push it to prod without more design input.

Patch-set: 1
Reviewer: Maryana <455@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1