Switching from php5 to php and removing unneeded variable in Makefile
There's no reason this should be specific to PHP version 5, plus
'php5' is not always registered as an alias to php. For example,
I'm running PHP 5.4.24, but 'php5' does nothing in my environment.

Removing MW_MF_INSTALL_PATH. We already have MW_INSTALL_PATH and
extensions are always in the extension dir, so I don't think we
need to create a separate variable for this.

Change-Id: I863ced57532cc4c41afdadc1e8aaa2e3eb605455