Update patch set 1
Patch Set 1:

Bmansurov: that's right. This is a backport. The branch wmf9 is the upcoming branch, which will be online on all wikipedias for 7 days starting this Wednesday. It's branch point is before the change, which was merged in master, so it still has the issue, which is already foxed in master. So, for some critical changes we backport them to the actual and upcoming version, to fix the problem there, toom

Patch-set: 1
Reviewer: Florianschmidtwelzow <890@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0