Icon: Remove confusing '0' argument to '.children()'
Follows-up 5c82ce1a9a.

.children() returns the child elements of the node(s) and
takes an optional CSS selector to filter these children.

Passing it a number has unexpected results. Normally this would
be casted to a string and used as selector and return an empty
jQuery collection and break the Icon class. However, as jQuery
only consideres non-falsey values, it instead returns all children.
(As if no argument was passed.)

It still worked because, in fact, no reduction to the first child
is needed at all.
* This Icon template only has one element.
* Getter methods like .attr(), .val() etc. can return only one
  value, that of the first element. This is a documented and
  established jQuery convention.

Change-Id: Id013465f0093e06795c978441099cc7fc2056fa9