Update patch set 6
Patch Set 6:

That works for me, it changes pages but it loads the last item from the previous page in the next page (i've changed the limit to 3)

Album here:
http://imgur.com/a/jiZVq#0

* First image full list
* Second, first page with limit changed to 3
* Third, second page with limit changed to 3
* Fourth, third page with limit changed to 3

Is this the behavior intended? I would expect the new page to be completely new items.

Not voting since I'm not sure if that is the intended behavior

Patch-set: 6
Reviewer: Jhernandez <2027@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0