Update patch set 8
Patch Set 8: Code-Review-1

This solves the problem, but is a fragile solution. We need to solve this the same way it is solved in core: mapping the legacy RestrictionLevels to their actual rights. In other words, the API needs to map sysop to 'editprotected' and autoconfirmed to 'editsemiprotected'. Then we should only be checking against the user rights, not the user groups at all.

Patch-set: 8
Label: Code-Review=-1