Update patch set 10
Patch Set 10:

After i talked to Bartosz (MatmaRex) and tested it on mw.org, desktop doesn't have such a feature. As a not logged-in, blocked user you still see the "edit" and "edit source" tab, instead of a "view source".

Nevertheless i will amend the change to give a better error message in the toast when the saving fails in order of a blocked user (actually only a "edit can not be saved" iirc).

Patch-set: 10
Label: Verified=0