Update patch set 9
Patch Set 9: Code-Review-1

(1 comment)

Lots of minor issues, but I guess it's a good start.

It would be nice if saving returned you to Special:Preferences. Would that be hard to do?

Also, the text on tablet looks tiny. Can we make the small font sizing mobile-only with a media query?

Patch-set: 9
Label: Code-Review=-1