Let PHP check, if the user can edit a page or not
Check if a page is editable by the current user in PHP. That includes
protection of articles and if the user has the correct right(s) to edit.

All Anonymous users should always have the same rights, still when they are blocked,
so there is no disadvantage of caching. The check, if the user is blocked from
editing the particular page, is limited to logged in users. If a not-
loggedin user is blocked and try to edit, he will get the error message after the failed
save api call.

Bug: T74877
Change-Id: I1ab70ac0f61238c39a43e6db63c75df9b5465819